requirements.txt
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
networkx>=2.6.0
matplotlib>=3.4.0
tqdm>=4.65.0
pydantic>=2.0.0
PyYAML>=6.0.0
requests>=2.31.0
huggingface_hub[cli]>=0.19.0
rich>=13.0.0
mypy>=1.0.0
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0
python-dotenv>=0.19.0
loguru>=0.6.0
structlog>=22.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
python-multipart>=0.0.5

# Development workflow
lightning>=2.1.0
lightning[extra]>=2.1.0  # Additional callbacks and utilities
wandb>=0.16.0
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
hydra-optuna-sweeper>=1.2.0
datasets>=2.15.0
torchmetrics>=1.2.0
pytorch-forecasting>=1.0.0
optuna<3.0.0,>=2.10.0
mlflow>=2.9.0
dvc>=3.30.0
psutil>=5.9.0
gputil>=1.4.0